+++ backend/src/main/java/com/bookbook/domain/rent/controller/RentController.java
@@ -34,7 +34,7 @@ public class RentController {
     @GetMapping("/{id}") // /rent/{id} 경로로 GET 요청을 처리
     @Operation(summary = " Rent 페이지 단건 조회")
     public RentResponseDto getRentPage(
-            @PathVariable int id, // 경로 변수로 전달된 id를 사용, 글 id
+            @PathVariable long id, // 경로 변수로 전달된 id를 사용, 글 id
             @AuthenticationPrincipal CustomOAuth2User customOAuth2User
     ){ 
         // 인증된 사용자 ID 가져오기 (비로그인 사용자는 null)
@@ -46,7 +46,7 @@ public class RentController {
     @PutMapping("/edit/{id}") // /rent/edit/{id} 경로로 PUT 요청을 처리
     @Operation(summary = "Rent 페이지 수정")
     public void editRentPage(
-            @PathVariable int id, // 경로 변수로 전달된 id를 사용, 글 id
+            @PathVariable long id, // 경로 변수로 전달된 id를 사용, 글 id
             @RequestBody @Valid RentRequestDto dto,
             @AuthenticationPrincipal CustomOAuth2User customOAuth2User
     ) {
