+++ backend/src/main/java/com/bookbook/domain/wishList/dto/WishListCreateRequestDto.kt
@@ -1,5 +1,5 @@
 package com.bookbook.domain.wishList.dto
 
 data class WishListCreateRequestDto(
-    val rentId: Int
+    val rentId: Long
 )
